Fix precision issue (#7214)

+++ bundles/org.openhab.binding.weathercompany/src/main/java/org/openhab/binding/weathercompany/internal/model/PwsObservationsDTO.java
@@ -46,7 +46,7 @@ public class PwsObservationsDTO {
         /*
          * The relative humidity of the air
          */
-        public Integer humidity;
+        public Double humidity;
 
         /*
          * Latitude of the PWS
